+++ src/lib/evaluator.h
@@ -2,6 +2,7 @@
 #define EVALUATOR_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "hashmap.h"
 #include "math.h"
@@ -11,6 +12,6 @@
 * @param: str: char*
 * @return: int
 */
-int evaluate(char* expression, hashmap* variables, bool* error_flag);
+int64_t evaluate(char* expression, hashmap* variables, bool* error_flag);
 
 #endif
