+++ src/lib/hashmap.h
@@ -3,12 +3,13 @@
 
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 #define P 61
 #define N 1000003
 
 typedef struct Hashmap{
-    int* container;
+    int64_t* container;
     int size;
 } hashmap;
 
@@ -16,8 +17,8 @@ hashmap* map_construct(int size);
 
 int hash(char* string);
 
-void map_assign(hashmap* map, char* variable, int value);
+void map_assign(hashmap* map, char* variable, int64_t value);
 
-int map_get(hashmap* map, char* variable);
+int64_t map_get(hashmap* map, char* variable);
 
 #endif
